+++ src/MinEdLauncher/Settings.fs
@@ -65,13 +65,18 @@ let parseArgs defaults (findCbLaunchDir: Platform -> Result<string,string>) (arg
             && [ Path.Combine("steamapps", "common", "Proton"); Path.Combine("Steam", "compatibilitytools.d", "Proton") ]
                |> List.exists (fun p -> argv.[0].Contains(p))
         let isNewProton() = // Proton >= 5.13 runs via steam linux runtime
-            argv.Length > 6 && argv.[0] <> null
+            argv.Length > 2 && argv.[0] <> null
             && argv.[0].Contains("SteamLinuxRuntime")
-        if isNewProton() then 
-            Some { EntryPoint = argv.[0]; Args = seq { argv.[1..4]; [| "python3" |]; argv.[5..6] } |> Array.concat }, Path.GetDirectoryName(argv.[7]) |> Some, argv.[7..]
+        if isNewProton() then
+            let doesntEndWith (value: string) (str: string) = not (str.EndsWith(value))
+            let runtimeArgs = argv.[1..] |> Array.takeWhile (doesntEndWith "proton")
+            let protonArgs = argv.[1..] |> Array.skipWhile (doesntEndWith "proton") |> Array.takeWhile (doesntEndWith "EDLaunch.exe")
+            let args = seq { runtimeArgs; [| "python3" |]; protonArgs } |> Array.concat
+            let launcherIndex = args.Length
+            Some { EntryPoint = argv.[0]; Args = args }, Path.GetDirectoryName(argv.[launcherIndex]) |> Some, argv.[launcherIndex..]
         else if isOldProton() then 
             Some { EntryPoint = "python3"; Args = argv.[..1] }, Path.GetDirectoryName(argv.[2]) |> Some, argv.[2..]
-        else if argv.Length > 0 && argv.[0] <> null && argv.[0].Contains("EDLaunch.exe", StringComparison.OrdinalIgnoreCase) then
+        else if argv.Length > 0 && argv.[0] <> null && argv.[0].EndsWith("EDLaunch.exe", StringComparison.OrdinalIgnoreCase) then
             None, Some (Path.GetDirectoryName(argv.[0])), argv.[1..]
         else
             None, None, argv
@@ -168,9 +173,9 @@ let parseConfig fileName =
         { config with Processes = processes } |> ConfigParseResult.Ok
     | Error error -> Error error
     
-let getSettings args fileConfig = task {
+let getSettings args appDir fileConfig = task {
     let findCbLaunchDir paths =
-        paths
+        appDir :: paths
         |> List.map Some
         |> List.append [ fileConfig.GameLocation ]
         |> List.choose id